# improve visiblity scope performance by EXIST