@oliverguenther 2022-10-31
This query is fixed on the way to 12.3 in this commit https://github.com/opf/openproject/commit/3bf706b153a5503677409526b033034c0cdac84a due to the removed distinct, as this requires the subselect to evaluate completely. Without it, the two modes are identical in my testing. 

I'm still fine with merging it as the resulting SQL is more explicit what we want, but it's a bit harder to read in Rails.