@ulferts 2022-10-31
@oliverguenther thanks for the digging!

I'll close this PR since it is indeed unusual to do this in Rails.